requirements.txt
flask
flask-wtf
flask-login
flask-restful
Flask-SQLAlchemy
python-dotenv
requests
qrcode
pillow
markupsafe
psycopg2-binary
wtforms
